remove uneeded utils file
